# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "Complete API for mechanic shop management with inventory tracking, JWT authentication, rate limiting, and caching"
  version: "1.0.0"
  contact:
    name: "Mechanic Shop API Support"
    email: "support@mechanicshop.com"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JWT Authorization header using the Bearer scheme.
      
      **Format:** `Bearer {jwt_token}`
      
      **Example:** `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      
      **How to obtain token:**
      1. Make a POST request to `/login` with valid credentials
      2. Copy the `access_token` from the response
      3. Add it to the Authorization header as `Bearer {access_token}`
      
      **Token contains:**
      - User identity and role information
      - Expiration time (typically 24 hours)
      - Permissions and access level
      
      **Usage:**
      - Include this header in requests to protected endpoints
      - Token will be validated on each request
      - Expired tokens will return 401 Unauthorized
  
  JWTAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JSON Web Token (JWT) authentication for API access.
      
      **Authentication Flow:**
      1. **Login:** POST credentials to `/login` endpoint
      2. **Token:** Receive JWT access token in response
      3. **Header:** Include token in Authorization header
      4. **Access:** Make authenticated requests to protected resources
      
      **Token Security:**
...
paths:
...
  /login:
    post:
      tags:
        - "Authentication"
      summary: "User authentication and token generation"
      description: |
        Universal login endpoint for all user types (customers, mechanics, admin).
        Authenticates user credentials and generates JWT token for API access.
        
        **Purpose:**
        - Central authentication endpoint
        - Token generation for protected resource access
        - Standardized login process across user types
        
        **Security Features:**
        - JWT token with 24-hour expiration
        - Secure password validation
        - Rate limiting protection
        
        **Usage:**
        1. Send credentials to this endpoint
        2. Receive JWT token in response
        3. Include token in Authorization header for protected endpoints
        4. Format: `Authorization: Bearer <token>`
      operationId: "generateToken"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: credentials
          description: "User login credentials for token generation"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Authentication successful - Token generated"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              message: "Authentication successful"
              status: "success"
              user:
                id: 1
                email: "user@example.com"
                type: "customer"
              timestamp: "2024-12-24T10:30:00Z"
        400:
          description: "Bad request - Invalid credentials format"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid request"
              message: "Email and password are required"
              timestamp: "2024-12-24T10:30:00Z"
        401:
          description: "Unauthorized - Invalid credentials"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Authentication failed"
              message: "Invalid email or password"
              timestamp: "2024-12-24T10:30:00Z"
        422:
          description: "Unprocessable entity - Validation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Validation error"
              message: "Invalid email format"
              timestamp: "2024-12-24T10:30:00Z"
        429:
          description: "Too many requests - Rate limit exceeded"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Rate limit exceeded"
              message: "Too many login attempts. Please try again later"
              timestamp: "2024-12-24T10:30:00Z"
...
  /customers/login:
    post:
      tags:
        - "Authentication"
      summary: "Customer login"
      description: |
        Authenticate a customer using email and password credentials. 
        Returns a JWT token for accessing protected endpoints.
        
        **Process:**
        1. Validates email format and password
        2. Checks credentials against database
        3. Generates JWT token with customer information
        4. Returns token and customer details
        
        **Security:** 
        - Password is validated securely
        - JWT token expires in 24 hours
        - No rate limiting applied to login endpoint
      operationId: "customerLogin"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: credentials
          description: "Customer login credentials (email and password)"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Authentication successful - JWT token generated"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Login successful"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              customer:
                id: 1
                first_name: "John"
                last_name: "Doe"
                email: "john.doe@test.com"
                phone_number: "555-1234"
        400:
          description: "Bad request - Invalid input format"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Validation failed"
              message: "Invalid email format"
        401:
          description: "Authentication failed - Invalid credentials"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Authentication failed"
              message: "Invalid email or password"
        422:
          description: "Unprocessable entity - Missing required fields"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Missing required fields"
              message: "Email and password are required"
  /customers/:
    get:
      tags:
        - "Customers"
      summary: "Get all customers"
      description: "Get all customers with pagination, sorting, and filtering"
      parameters:
        - in: query
          name: page
          type: integer
          description: "Page number"
          default: 1
        - in: query
          name: per_page
          type: integer
          description: "Items per page (max 100)"
          default: 10
        - in: query
          name: sort_by
          type: string
          description: "Sort field"
          enum: ["id", "first_name", "last_name", "email"]
        - in: query
          name: order
          type: string
          description: "Sort order"
          enum: ["asc", "desc"]
        - in: query
          name: search
          type: string
          description: "Search term for name or email"
        - in: query
          name: email
          type: string
          description: "Filter by email"
      responses:
        200:
          description: "Customers retrieved successfully"
    post:
      tags:
        - "Customers"
      summary: "Create customer"
      description: "Create a new customer (Rate Limited: 5 per minute)"
      parameters:
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerCreate"
      responses:
        201:
          description: "Customer created successfully"
        400:
          description: "Validation error"
        429:
          description: "Rate limit exceeded"
  /customers/{id}:
    get:
      tags:
        - "Customers"
      summary: "Get customer by ID"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Customer retrieved successfully"
        404:
          description: "Customer not found"
    put:
      tags:
        - "Customers"
      summary: "Update customer"
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: customer
          schema:
            $ref: "#/definitions/CustomerUpdate"
      responses:
        200:
          description: "Customer updated successfully"
        403:
          description: "Unauthorized"
        404:
          description: "Customer not found"
...
definitions:
  Customer:
    type: object
    properties:
      id:
        type: integer
        example: 1
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      email:
        type: string
        format: email
        example: "john.doe@test.com"
      phone_number:
        type: string
        example: "555-1234"
  CustomerCreate:
    type: object
    required:
      - first_name
      - last_name
      - email
      - password
    properties:
      first_name:
        type: string
        minLength: 1
        maxLength: 100
        example: "John"
      last_name:
        type: string
        minLength: 1
        maxLength: 100
        example: "Doe"
      email:
        type: string
        format: email
        example: "john.doe@test.com"
      phone_number:
        type: string
        maxLength: 20
        example: "555-1234"
      password:
        type: string
        minLength: 6
        example: "securepass123"
  CustomerUpdate:
    type: object
    properties:
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      phone_number:
        type: string
        example: "555-1234"
...
  LoginCredentials:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        description: "Customer's registered email address"
        example: "john.doe@test.com"
        minLength: 5
        maxLength: 255
      password:
        type: string
        description: "Customer's password (minimum 6 characters)"
        example: "securepass123"
        minLength: 6
        maxLength: 255
    example:
      email: "john.doe@test.com"
      password: "securepass123"
  LoginResponse:
    type: object
    required:
      - message
      - token
      - status
    properties:
      token:
        type: string
        description: "Authentication token used by clients to authenticate themselves in subsequent requests to protected endpoints"
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        minLength: 1
      message:
        type: string
        description: "Informative message indicating the success or failure of the login operation to help clients understand the outcome"
        example: "Login successful"
      status:
        type: string
        description: "Status of the operation indicating success or failure"
        example: "success"
        enum: ["success", "error"]
      customer:
        $ref: "#/definitions/Customer"
        description: "Customer information (excluding password for security)"
      timestamp:
        type: string
        format: date-time
        description: "When the login occurred"
        example: "2024-12-24T10:30:00Z"
    example:
      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      message: "Login successful"
      status: "success"
      customer:
        id: 1
        first_name: "John"
        last_name: "Doe"
        email: "john.doe@test.com"
        phone_number: "555-1234"
      timestamp: "2024-12-24T10:30:00Z"
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        description: "Error category or type"
        example: "Authentication failed"
      message:
        type: string
        description: "Detailed error message"
        example: "Invalid email or password"
      timestamp:
        type: string
        format: date-time
        description: "When the error occurred"
        example: "2024-12-24T10:30:00Z"
    example:
      error: "Authentication failed"
      message: "Invalid email or password"
      timestamp: "2024-12-24T10:30:00Z"

  # Member Management Definitions
...